+++ examples/lesson2_player_movement.py
@@ -1,4 +1,4 @@
-from kaplay import *
+from kaypy import *
 
 kaplay(width=800, height=600, background=[0, 0, 0])