+++ examples/lesson7_scenes.py
@@ -1,4 +1,4 @@
-from kaplay import *
+from kaypy import *
 
 kaplay(width=800, height=600)
 setBackground(0, 0, 0)